// MercaditoTec/src/app/models/job-offer-interface.ts
export interface JobOfferInterface {
    idOfertaLaboral: number;
    idEmpleador: number;
    nombrePuesto: string;
    responsabilidades: string;
    requerimientos: string;
    idCarrera: number;
    idUbicacion: number;
    jornadaLaboral: string;
    link: string;
    fechaPublicacion: string;
}